Rework for final presentation

+++ server/api/route/projects.py
@@ -7,6 +7,7 @@ from api.services.guards import authorization_guard
 from api.model.project import Project
 from sqlalchemy.exc import IntegrityError
 from api.model.user import User
+from api.model.task import Task
 from flask import g
 
 
@@ -15,20 +16,21 @@ project_api = Blueprint("project_api", __name__)
 projects_schema = ProjectsSchema()
 
 
-@project_api.get("/projects", endpoint="getProjectList")
-@authorization_guard
-def getProjectList():
-    userID = g.access_token["sub"]
-    projects = Project.query.join(User).filter(User.userID == userID).all()
+#TODO: Return object itself
+
+@project_api.get("/<string:username>/projects", endpoint="getProjectList")
+def getProjectList(username):
+    
+    projects = Project.query.join(User).filter(User.userID == username).all()
     result = projects_schema.dump(projects, many=True)
-    return {"projects": result}
+    return result, 200
 
 
-@project_api.post("/projects", endpoint="createProject")
-@authorization_guard
-def createProject():
+@project_api.post("/<string:userID>/projects", endpoint="createProject")
+def createProject(userID):
 
     json_data = request.get_json()
+    print(json_data)
 
     if not json_data:
         return {"message": "Empty request"}, 400
@@ -41,29 +43,27 @@ def createProject():
 
     try:
         new_project = Project(**data)
-        new_project.userID = g.access_token["sub"]
+        new_project.userID = userID
         db.session.add(new_project)
         db.session.commit()
     except IntegrityError:
         return {"message": "Error creating project"}, 400
 
     result = projects_schema.dump(new_project)
-    return {"message": "New project created.", "project": result}, 201
+    return result, 201
 
 
-@project_api.get("projects/<uuid:project_id>", endpoint="getProject")
-@authorization_guard
-def getProject(project_id):
-    project = Project.query.get(project_id)
+@project_api.get("/<userID>/projects/<uuid:projectID>", endpoint="getProject")
+def getProject(userID, projectID):
+    project = Project.query.get(projectID)
     if project is None:
         return {"message": "Project does not exist."}, 404
     result = projects_schema.dump(project)
-    return {"project": result}, 200
+    return result, 200
 
 
-@project_api.put("projects/<uuid:project_id>", endpoint="editProject")
-@authorization_guard
-def editProject(project_id):
+@project_api.put("/<userID>/projects/<uuid:projectID>", endpoint="editProject")
+def editProject(userID, projectID):
     json_data = request.get_json()
     if not json_data:
         return {"message": "Empty request"}, 400
@@ -72,28 +72,34 @@ def editProject(project_id):
     except ValidationError as err:
         return err.messages, 422
 
+    if updated_data["projectID"] != projectID:
+        return {"message":"projectID does not match url"}, 403
     # Have to use query object to handle case where creation is necessary
     try:
-        Project.query.filter(Project.projectID == project_id).update(
+        Project.query.filter(Project.projectID == projectID).update(
             values=updated_data
         )
         db.session.commit()
     except Exception as e:
         return {"message": e}
 
-    project = Project.query.get(project_id)
+    project = Project.query.get(projectID)
     result = projects_schema.dump(project)
 
-    return {"message": "Project updated.", "project": result}, 200
+    return result, 200
 
 
-@project_api.delete("projects/<uuid:project_id>")
-@authorization_guard
-def deleteProject(project_id):
-    project = Project.query.get(project_id)
+@project_api.delete("/<userID>/projects/<uuid:projectID>")
+def deleteProject(userID, projectID):
+    project = Project.query.get(projectID)
     if project is None:
         return {"message": "Project does not exist"}, 404
 
+    if project.userID != userID:
+        return {"message":"User does not own project"}, 403
     db.session.delete(project)
     db.session.commit()
     return {}, 204
+
+
+
